+++ StripePaymentSheet/StripePaymentSheetTests/PaymentSheet/STPFixtures+PaymentSheet.swift
@@ -448,6 +448,7 @@ extension PaymentSheet.Appearance {
 
         // Customize the primary button
         var primaryButton = PaymentSheet.Appearance.PrimaryButton()
+        primaryButton.height = 50
         primaryButton.cornerRadius = 8
 
         appearance.font = font